refactor: extract Config dataclass from config_loader into _config_class module
Moves the Config dataclass to _config_class.py, keeping singleton
management and lazy accessors in config_loader.py as the entry point.
