+++ src/quodeq/shared/config_loader.py
@@ -1,63 +1,16 @@
 """Centralized configuration loading for the Quodeq package.
 
-Provides the :class:`Config` dataclass and lazy-loaded default constants
-read from ``defaults.json``.
+Provides the :class:`Config` dataclass (via ``_config_class``) and
+lazy-loaded default constants read from ``defaults.json``.
 """
 from __future__ import annotations
 
 import threading
-from contextlib import contextmanager
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Any, Iterator
+
+from quodeq.shared._config_class import Config  # noqa: F401
 
 _DEFAULTS_PATH = Path(__file__).resolve().parent / "defaults.json"
-
-
-@dataclass
-class Config:
-    """Centralized configuration holder loaded from defaults.json.
-
-    Replaces raw module-level mutable dict with a testable object that
-    supports safe overrides via the :meth:`override` context manager.
-    """
-
-    _data: dict[str, Any] = field(default_factory=dict, init=False)
-
-    def __getitem__(self, key: str) -> Any:
-        return self._data[key]
-
-    def get(self, key: str, default: Any = None) -> Any:
-        return self._data.get(key, default)
-
-    def update(self, **overrides: Any) -> None:
-        self._data.update(overrides)
-
-    @contextmanager
-    def override(self, **overrides: Any) -> Iterator[None]:
-        """Temporarily override config values; restores originals on exit."""
-        saved = {k: self._data[k] for k in overrides if k in self._data}
-        removed = {k for k in overrides if k not in self._data}
-        self._data.update(overrides)
-        try:
-            yield
-        finally:
-            self._data.update(saved)
-            for k in removed:
-                self._data.pop(k, None)
-
-    @classmethod
-    def from_file(cls, path: Path) -> Config:
-        obj = cls()
-        try:
-            from quodeq.shared.utils import read_json
-            obj._data = read_json(path)
-        except (OSError, ValueError) as exc:
-            import logging
-            logging.getLogger(__name__).warning("Failed to load config from %s: %s", path, exc)
-            obj._data = {}
-        return obj
-
 
 _config_lock = threading.Lock()
 # Singleton instance — access via _get_config(override) for DI; reset via _reset_config_for_testing().
